+++ src/course_project_suite/__init__.py
@@ -1,2 +1,2 @@
 __all__ = ["common"]
-__version__ = "4.3.0"
+__version__ = "4.4.0"